requirements.txt
# for install every dependencies, use
# >>> pip install -r requirements.txt
python-decouple==3.8
pyttsx3==2.99
SpeechRecognition==3.14.3
